refactor: separar rutas y controladores
