+++ backend-pim/controllers/usersController.js
@@ -1,0 +1,76 @@
+import prisma from '../prismaClient.js';
+import bcrypt from 'bcryptjs';
+import jwt from 'jsonwebtoken';
+
+const SECRET = process.env.JWT_SECRET;
+
+export async function getUser(req, res) {
+  const userId = parseInt(req.params.id);
+
+  if (isNaN(userId)) {
+    return res.status(400).json({ error: 'ID no vàlid' });
+  }
+
+  try {
+    const user = await prisma.user.findUnique({
+      where: { id: userId },
+      select: {
+        id: true,
+        name: true,
+        email: true,
+      },
+    });
+
+    if (!user) {
+      return res.status(404).json({ error: 'Usuari no trobat' });
+    }
+
+    res.json(user);
+  } catch (err) {
+    console.error('Error obtenint usuari:', err);
+    res.status(500).json({ error: 'Error al carregar usuari' });
+  }
+}
+
+export async function registerUser(req, res) {
+  const { name, email, password } = req.body;
+  if (!name || !email || !password) {
+    return res.status(400).json({ error: 'Nom, email i password són obligatoris' });
+  }
+
+  try {
+    const hashedPassword = await bcrypt.hash(password, 10);
+    const user = await prisma.user.create({
+      data: { name, email, password: hashedPassword },
+    });
+    res.status(201).json({ id: user.id, name: user.name, email: user.email });
+  } catch (err) {
+    console.error('Error registrant usuari:', err);
+    res.status(500).json({ error: 'Error al registrar l\'usuari' });
+  }
+}
+
+export async function loginUser(req, res) {
+  const { email, password } = req.body;
+  if (!email || !password) {
+    return res.status(400).json({ error: 'Email i password són obligatoris' });
+  }
+
+  try {
+    const user = await prisma.user.findUnique({ where: { email } });
+    if (!user) {
+      return res.status(404).json({ error: 'Usuari no trobat' });
+    }
+
+    const isValid = await bcrypt.compare(password, user.password);
+    if (!isValid) {
+      return res.status(401).json({ error: 'Password incorrecte' });
+    }
+
+    const token = jwt.sign({ userId: user.id }, SECRET, { expiresIn: '7d' });
+    res.json({ token });
+  } catch (err) {
+    console.error('Error login:', err);
+    res.status(500).json({ error: 'Error al iniciar sessió' });
+  }
+}
