+++ backend-pim/middleware/auth.js
@@ -1,0 +1,15 @@
+import jwt from 'jsonwebtoken';
+
+const SECRET = process.env.JWT_SECRET;
+
+export function authenticateToken(req, res, next) {
+  const authHeader = req.headers['authorization'];
+  const token = authHeader && authHeader.split(' ')[1];
+  if (!token) return res.sendStatus(401);
+
+  jwt.verify(token, SECRET, (err, user) => {
+    if (err) return res.sendStatus(403);
+    req.userId = user.userId;
+    next();
+  });
+}
