+++ backend-pim/middleware/upload.js
@@ -1,0 +1,21 @@
+import multer from 'multer';
+import path from 'path';
+import fs from 'fs';
+
+export const uploadDir = './uploads';
+if (!fs.existsSync(uploadDir)) {
+  fs.mkdirSync(uploadDir);
+}
+
+const storage = multer.diskStorage({
+  destination: function (req, file, cb) {
+    cb(null, uploadDir);
+  },
+  filename: function (req, file, cb) {
+    const uniqueName = Date.now() + '-' + Math.round(Math.random() * 1e9);
+    cb(null, uniqueName + path.extname(file.originalname));
+  }
+});
+
+export const upload = multer({ storage });
+export const uploadImport = multer({ dest: uploadDir });
