+++ backend-pim/prismaClient.js
@@ -1,0 +1,3 @@
+import { PrismaClient } from './generated/prisma/index.js';
+const prisma = new PrismaClient();
+export default prisma;
